Combined all files into main.cpp
