--- student.cpp
@@ -1,3 +1,0 @@
-//implementation of the student class for the maintenance roll system.
-#include <iostream>
-#include "student.h"
