+++ students.csv
@@ -1,0 +1,5 @@
+Igor Amidzic,1234-1234,igoramidzic@usf.edu,100,100,100
+Oleh Tielushko,234-1234,igoramidzic@usf.edu,100,100,100
+Yakira Quemba,1234-1234,yakiraq@usf.edu,100,100,100
+Thinh Phung,1234-1234,thinhphung@usf.edu,100,100,100
+Jack Dew,1234-0000,jdew@usf.edu,100,100,100
